+++ src/packages/eslint-config-kiwicom/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# 2.8.0
 - Add `FormData` to the globals
 - Rule `import/no-unresolved` now ignores Relay artifacts set by `artifactDirectory`
 - Upgrade `eslint-plugin-react-hooks` to the latest stable version (^1.0.1)